+++ chat-widget/src/components/webchatcontainerstateful/interfaces/IWebChatProps.ts
@@ -43,4 +43,5 @@ export interface IWebChatProps {
     //                 I.e. not using composition mode (or <Composer>).
     className?: string;
     role?: "complementary" | "contentinfo" | "form" | "main" | "region";
+    receivedMessageAnchorColor?: string;
 }